// src/app/models/sales/StocklineListSalesFilter.ts
export class StocklineListSalesFilter {
    stockLineId: number;
    itemMasterId: number
    partNumber: string;
    partDescription: string;
    itemGroup: string;
    itemCategory: string;
    stocklineNumber: string;
    serialNumber: string;
    controlIdNumber: string;
    controlNumber: string;
    condition: string;
    quantityOnHand: number;
    quantityAvailable: number;
    glAccountName: number;
    glAccount: string;
    unitSalesPrice: number;
    saleUnitCost: number;
    createdDate: string;
    isActive: boolean;
    conditionId: number;
}